refactor(runtime): split dispatch.rs into cohesive modules (§7-8)
Split the 2169-line dispatch.rs (over the 500-line limit) into six
cohesive modules, all under 500 lines:

- dispatch.rs (219)               - DispatchShape, pending-error/visible-
                                    params helpers, grammar-token dispatch
- dispatch_candidates.rs (498)    - candidate selection/specificity/type-distance
- dispatch_resolve.rs (474)       - resolve_function_with_* + candidate listing
- dispatch_proto.rs (357)         - proto lookup/resolution/method body
- dispatch_proto_call.rs (314)    - call_proto_dispatch + candidate rebuild
- dispatch_proto_rewrite.rs (324) - {*} proto-dispatch AST rewrite + env restore

All methods are impl Interpreter, so the split is visibility-only. Nine
cross-module private helpers are bumped to pub(super), and the
is_coercion_constraint import is relocated to the candidates module that
uses it. The fn multiset is byte-identical. No behavior change.
