+++ src/runtime/mod.rs
@@ -171,6 +171,11 @@ mod class_introspection;
 pub(crate) mod deprecation;
 pub(crate) mod did_you_mean;
 mod dispatch;
+mod dispatch_candidates;
+mod dispatch_proto;
+mod dispatch_proto_call;
+mod dispatch_proto_rewrite;
+mod dispatch_resolve;
 mod eval_check;
 mod handle;
 mod handle_io;
